+++ example/time.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: avoid_print, prefer_const_constructors
+// ignore_for_file: avoid_print, prefer_const_constructors, unused_local_variable
 
 import 'package:date_time_iso/date_time_iso.dart';