+++ test/date_test.dart
@@ -130,9 +130,9 @@ void main() {
 
     then('is this week', () {
       if (today.weekday == DateTime.sunday) {
-        today.subDays(1).isThisIsoWeek.should.beTrue();
+        today.subDays(1).isThisWeek.should.beTrue();
       } else {
-        today.addDays(1).isThisIsoWeek.should.beTrue();
+        today.addDays(1).isThisWeek.should.beTrue();
       }
 
       if (today.weekday == DateTime.saturday) {